+++ crates/async-dropper-simple/Cargo.toml
@@ -33,7 +33,6 @@ rustc_version = "0.4.0"
 
 [dev-dependencies]
 async-std = { workspace = true, features = ['attributes'] }
-macros_internal = { path = "../macros_internal" }
 tokio = { workspace = true, features = [
   "time",
   "macros",
